# ApiDocs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HKL Planet Web API",
    "contact": {
      "name": "Marcelo de Oliveira Francisco",
      "email": "marcdfranc@gmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "v1.0.0",
    "x-logo": {
      "url": "assets/logo.png",
      "backgroundColor": "#FFFFFF",
      "altText": "Marcelo Dotnet Developer"
    }
  },
  "servers": [
    {
      "url": "",
      "description": "PlanetApi"
    }
  ],
  "paths": {
    "/api/Planet": {
      "post": {
        "tags": [
          "Planet"
        ],
        "summary": "Create Planet",
        "description": "Create a new Planet registry",
        "operationId": "CreatePlanet",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/PlanetData"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanetData"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanetData"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlanetData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Planet"
        ],
        "summary": "Get Planets",
        "description": "returns a paged array of Planets",
        "operationId": "GetPlanets",
        "parameters": [
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Pagination": {
                "schema": {
                  "$ref": "#/components/schemas/Pagination"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanetResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Planet/{planetId}": {
      "get": {
        "tags": [
          "Planet"
        ],
        "summary": "Detail Planet",
        "description": "Get Planet detail by uuid",
        "operationId": "DetailPlanet",
        "parameters": [
          {
            "name": "planetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Planet"
        ],
        "summary": "Edit Planet",
        "description": "Edit a Planet by uuid",
        "operationId": "EditPlanet",
        "parameters": [
          {
            "name": "planetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/PlanetData"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanetData"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanetData"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlanetData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Planet"
        ],
        "summary": "Delete Planet",
        "description": "Delete a Planet by uuid",
        "operationId": "DeletePlanet",
        "parameters": [
          {
            "name": "planetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Pagination": {
        "type": "object",
        "properties": {
          "currentPage": {
            "type": "integer",
            "format": "int32"
          },
          "itemsPerPage": {
            "type": "integer",
            "format": "int32"
          },
          "totalItems": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PlanetData": {
        "required": [
          "circular",
          "diameter",
          "distanceFromTheSun",
          "imageIconPath",
          "imagePath",
          "mass",
          "name",
          "position"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 30,
            "minLength": 1,
            "type": "string"
          },
          "distanceFromTheSun": {
            "type": "number",
            "format": "double"
          },
          "mass": {
            "type": "number",
            "format": "double"
          },
          "diameter": {
            "type": "number",
            "format": "double"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "imagePath": {
            "minLength": 1,
            "type": "string"
          },
          "imageIconPath": {
            "minLength": 1,
            "type": "string"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "circular": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "PlanetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "decimalDistanceFromTheSun": {
            "type": "number",
            "format": "double"
          },
          "distanceFromTheSun": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "decimalMass": {
            "type": "number",
            "format": "double"
          },
          "mass": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "diameter": {
            "type": "number",
            "format": "double"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "imagePath": {
            "type": "string",
            "nullable": true
          },
          "imageIconPath": {
            "type": "string",
            "nullable": true
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "circular": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "x-api-key Authorization header using x-api-key scheme. Example \"x-api-key: {API_KEY}\"",
        "name": "x-api-key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}